+++ draft-packages/illustration/KaizenDraft/Illustration/index.ts
@@ -1,0 +1,2 @@
+export * from "./Spot"
+export * from "./Scene"