+++ packages/component-library/draft/Kaizen/ZenNavigationBar/NavigationBar.tsx
@@ -1,37 +1,38 @@
 import * as React from "react"
 
-import { ZenControlledOffCanvas } from "@kaizen/component-library/draft/Kaizen/ZenOffCanvas"
 import classNames from "classnames"
 import Media from "react-media"
 import uuid from "uuid/v4"
-import {
-  LocalBadge,
-  namedBadge,
-  ProductionBadge,
-  StagingBadge,
-  TestBadge,
-} from "./components/Badge"
+import { ZenControlledOffCanvas } from "../ZenOffCanvas"
+import Badge from "./components/Badge"
 import Link from "./components/Link"
 import Menu from "./components/Menu"
-import { Navigation, NavigationItem } from "./types"
+import { NavBarContext } from "./context"
+import { Navigation, NavigationChange, NavigationItem } from "./types"
 
 const styles = require("./NavigationBar.module.scss")
 
 type Props = {
   environment?: "production" | "staging" | "test" | "local"
   loading?: boolean
-  colorScheme?: "cultureamp" | "kaizen" | "content"
+  colorScheme?: ColorScheme
   badgeHref?: string
+  onNavigationChange: NavigationChange
   headerComponent?: {
     desktop: React.ReactNode
     mobile: React.ReactNode
   }
   footerComponent?: React.ReactNode
   children?: Navigation
-  mobileMaxWidth?: number
 }
 
-export default class NavigationBar extends React.Component<Props> {
+type State = {
+  mobileKey: number
+}
+
+export type ColorScheme = "cultureamp" | "kaizen" | "content"
+
+export default class NavigationBar extends React.Component<Props, State> {
   static displayName = "NavigationBar"
   static Link = Link
   static Menu = Menu
@@ -40,7 +41,12 @@ export default class NavigationBar extends React.Component<Props> {
     loading: false,
     colorScheme: "cultureamp",
     badgeHref: "/",
-    mobileMaxWidth: 767,
+    onNavigationChange: () => null,
+  }
+
+  constructor(props) {
+    super(props)
+    this.state = { mobileKey: 1 }
   }
 
   render() {
@@ -48,38 +54,57 @@ export default class NavigationBar extends React.Component<Props> {
       children,
       colorScheme = "cultureamp",
       headerComponent,
-      mobileMaxWidth,
+      onNavigationChange,
     } = this.props
 
     return (
-      <Media query={`(max-width: ${mobileMaxWidth}px)`}>
-        {(matches: boolean) =>
-          matches ? (
-            <ZenControlledOffCanvas
-              headerComponent={
-                headerComponent ? headerComponent.mobile : this.renderBadge()
-              }
-              footerComponent={this.props.footerComponent}
-              links={children}
-              heading="Menu"
-              menuId="menu"
-            />
-          ) : (
-            <header
-              className={classNames(styles.navigationBar, styles[colorScheme])}
-            >
-              {headerComponent ? (
-                <span className={styles.headerSlot}>
-                  {headerComponent.desktop}
-                </span>
-              ) : (
-                this.renderBadge()
-              )}
-              {this.renderNav(children)}
-            </header>
-          )
-        }
-      </Media>
+      <NavBarContext.Provider
+        value={{
+          handleNavigationChange: event => {
+            const navigationHref = event.currentTarget.getAttribute("href")
+            if (navigationHref && navigationHref !== "#") {
+              this.setState({
+                mobileKey: this.state.mobileKey + 1,
+              })
+              onNavigationChange(event)
+            }
+          },
+          hasExtendedNavigation: !!children?.secondary?.length,
+        }}
+      >
+        <Media query={`(max-width: ${styles.caBreakpointMobileMax})`}>
+          {(matches: boolean) =>
+            matches ? (
+              <ZenControlledOffCanvas
+                key={this.state.mobileKey}
+                headerComponent={this.renderBadge()}
+                footerComponent={this.props.footerComponent}
+                productSwitcher={headerComponent && headerComponent.mobile}
+                links={children}
+                heading="Menu"
+                menuId="menu"
+                colorScheme={colorScheme}
+              />
+            ) : (
+              <header
+                className={classNames(
+                  styles.navigationBar,
+                  styles[colorScheme]
+                )}
+              >
+                {headerComponent ? (
+                  <span className={styles.headerSlot}>
+                    {headerComponent.desktop}
+                  </span>
+                ) : (
+                  this.renderBadge()
+                )}
+                {this.renderNav(children)}
+              </header>
+            )
+          }
+        </Media>
+      </NavBarContext.Provider>
     )
   }
 
@@ -115,7 +140,12 @@ export default class NavigationBar extends React.Component<Props> {
     const isFinal = section === "final"
     const linkWithProps = {
       ...link,
-      props: { ...linkProps, opaque: isFinal, small: isFinal },
+      props: {
+        ...linkProps,
+        opaque: isFinal,
+        small: isFinal,
+        content: this.props.colorScheme === "content",
+      },
     }
     const key = "href" in linkProps ? linkProps.href : linkProps.heading
 
@@ -133,20 +163,10 @@ export default class NavigationBar extends React.Component<Props> {
 
   renderBadge() {
     const {
-      environment = "production",
       loading = false,
       badgeHref = "/",
       colorScheme = "kaizen",
     } = this.props
-
-    const badges = {
-      production: ProductionBadge,
-      staging: StagingBadge,
-      test: TestBadge,
-      local: LocalBadge,
-    }
-
-    const Badge = badges[environment] || namedBadge(environment)
 
     return (
       <Badge loading={loading} href={badgeHref} colorScheme={colorScheme} />