+++ packages/component-library/draft/Kaizen/ZenOffCanvas/ZenOffCanvas.tsx
@@ -1,6 +1,7 @@
 import * as React from "react"
 
 import classNames from "classnames"
+import { ColorScheme } from "../ZenNavigationBar/NavigationBar"
 import Header from "./components/Header"
 import Menu from "./components/Menu"
 
@@ -11,7 +12,9 @@ type Props = {
   heading: string
   headerComponent: React.ReactNode
   footerComponent?: React.ReactNode
+  productSwitcher?: React.ReactNode
   menuId: string
+  colorScheme?: ColorScheme
 }
 
 type State = {
@@ -33,6 +36,7 @@ export const OffCanvasContext = React.createContext<OffCanvasContextProps>({
 export class ZenOffCanvas extends React.Component<Props> {
   static defaultProps = {
     withTrigger: false,
+    colorScheme: "cultureamp",
   }
 
   render() {
@@ -42,6 +46,8 @@ export class ZenOffCanvas extends React.Component<Props> {
       heading,
       links,
       footerComponent,
+      productSwitcher,
+      colorScheme,
     } = this.props
 
     return (
@@ -56,14 +62,28 @@ export class ZenOffCanvas extends React.Component<Props> {
               onClose={resetVisibleMenus}
               leftComponent={headerComponent}
               heading={heading}
+              colorScheme={colorScheme}
             />
-            <nav className={styles.links}>
-              {links &&
-                Object.keys(links).map(section => (
-                  <Menu section={section} link={links[section]} />
-                ))}
-            </nav>
-            {footerComponent}
+            <div
+              className={classNames(styles.contentContainer, {
+                [styles.hasFooter]: !!footerComponent,
+              })}
+            >
+              {productSwitcher && productSwitcher}
+              <nav className={styles.links}>
+                {links &&
+                  Object.keys(links).map(section => (
+                    <Menu
+                      key={section}
+                      section={section}
+                      link={links[section]}
+                    />
+                  ))}
+              </nav>
+            </div>
+            {footerComponent && (
+              <div className={styles.footerComponent}>{footerComponent}</div>
+            )}
           </div>
         )}
       </OffCanvasContext.Consumer>