+++ packages/component-library/draft/Kaizen/ZenOffCanvas/components/Header.tsx
@@ -1,5 +1,7 @@
 import * as React from "react"
 
+import classNames from "classnames"
+import { ColorScheme } from "../../ZenNavigationBar/NavigationBar"
 const closeIcon = require("@kaizen/component-library/icons/close.icon.svg")
   .default
 import IconButton from "@kaizen/component-library/components/Button/IconButton"
@@ -10,13 +12,28 @@ type Props = {
   leftComponent: React.ReactNode
   onClose: (e: MouseEvent) => void
   heading: string
+  colorScheme?: ColorScheme
 }
 
-const Header = ({ leftComponent, onClose, heading }: Props) => (
-  <div className={styles.root}>
+const Header = ({
+  leftComponent,
+  onClose,
+  heading,
+  colorScheme = "cultureamp",
+}: Props) => (
+  <div
+    className={classNames(styles.root, {
+      [styles.content]: colorScheme === "content",
+    })}
+  >
     {leftComponent}
     <span className={styles.heading}>{heading}</span>
-    <IconButton label="Close" icon={closeIcon} onClick={onClose} reversed />
+    <IconButton
+      label="Close"
+      icon={closeIcon}
+      onClick={onClose}
+      reversed={colorScheme !== "content"}
+    />
   </div>
 )