+++ storybook/main.ts
@@ -2,6 +2,12 @@ module.exports = {
   addons: ["@storybook/addon-docs"],
   stories: [
     "../packages/component-library/stories/*.stories.tsx",
-    "../draft-packages/**/*.stories.tsx",
+    /**
+     * These stories are separated from their corresponding packages as
+     * the packages contained compiled js in their node_modules and that
+     * made storybook unhappy
+     */
+    "../draft-packages/stories/*.stories.tsx",
+    "../legacy-packages/**/*.stories.tsx",
   ],
 }